я сделал отрисовку кривой Бизе! Ура товарищи! Празднуем!

+++ СomputerGraphics/src/main/java/Line2D.java
@@ -2,12 +2,12 @@
 public class Line2D {
 
 	
-	public static double sign (double x) {
+	public static int sign (int x) {
 		return (x > 0) ? 1 : (x < 0) ? -1 : 0;
 	}
 	
-	public static void AddLineSigmentOnDisplayBresenham(double xstart, double ystart, double xend, double yend) {
-		double x, y, dx, dy, incx, incy, pdx, pdy, es, el, err;
+	public static void AddLineSigmentOnDisplayBresenham(int xstart, int ystart, int xend, int yend) {
+		int x, y, dx, dy, incx, incy, pdx, pdy, es, el, err;
 
 
 		dx = xend - xstart;
